Update Makefile to enforce fresh collectstatic action
Issue: The Django `collectstatic` feature will *not* overwrite a file that already exists in the static files directory by default. If it finds a new, uncollected file it will be copied over. However, if it finds a new version of an existing file, it will ignore the new file in favor of keeping the old. Due to the way the code is now built and shared between containers, there needs to be a forced collection of the static files upon deployment to ensure the latest code is being picked up.

Solution: Adding the `--clear` option to the collectstatic command will force django to delete the contents of `STATICFILESDIR` before executing the collectstatic command, eliminating the problem.