Question about config code
I notice that we are splitting out a common.env file for the shared tags.

Isn't it easier to simply have one .env file with ALL the settings per $ENV HOST (which would match the sample.env file above) and eliminate maintaining the common.env file? I know it is not 100% DRY that way, but this is very short config and seems inefficient (opening a second file for 2 settings values) at this scale.

If common.env is going to be maintained, what is the intended practice for which of these sample.env values should go into that file instead of the $ENV.env file? There are enough deployments building up that we need to align on a consistent best practice for configuration.